Move test_text.py to test_text_eager.py
